First step solely responsible for all image uploads
